refactor
- refactor source into a more manageable structure.
- refactor tests to match new code structure
- move share constants to own file
