+++ src/_constants.js
@@ -1,0 +1,29 @@
+export const COINS = ['CP', 'SP', 'EP', 'GP', 'PP'];
+
+export const DENOMINATIONS = {
+  CP: {
+    name: 'copper',
+    abrv: 'CP',
+    copperValue: 1,
+  },
+  SP: {
+    name: 'silver',
+    abrv: 'SP',
+    copperValue: 10,
+  },
+  EP: {
+    name: 'electrum',
+    abrv: 'EP',
+    copperValue: 50,
+  },
+  GP: {
+    name: 'gold',
+    abrv: 'GP',
+    copperValue: 100,
+  },
+  PP: {
+    name: 'platinum',
+    abrv: 'PP',
+    copperValue: 1000,
+  },
+};
